+++ lib/logstash/filters/jdbc/basic_database.rb
@@ -1,3 +1,5 @@
+# encoding: utf-8
+require "fileutils"
 require "sequel"
 require "sequel/adapters/jdbc"
 require "java"
@@ -49,6 +51,8 @@ module LogStash module Filters module Jdbc
       rescue *CONNECTION_ERRORS => err
         # we do not raise an error when there is a connection error, we hope that the connection works next time
         logger.error(err_message, :exception => err.message, :backtrace => err.backtrace.take(8))
+      else
+        raise "::Sequel.connect returned a nil db instance, connection_string: #{@connection_string}, options: #{@options_hash.inspect}" if @db.nil?
       end
     end
 
@@ -80,7 +84,12 @@ module LogStash module Filters module Jdbc
 
     def verify_connection(connection_string, driver_class, driver_library, user, password)
       begin
-        require driver_library if driver_library
+        if driver_library
+          class_loader = java.lang.ClassLoader.getSystemClassLoader().to_java(java.net.URLClassLoader)
+          driver_library.split(",").each do |driver_path|
+            make_driver_path_loadable(class_loader, driver_path.strip)
+          end
+        end
       rescue LoadError => e
         msg = "The driver library cannot be loaded. The system error was: '#{e.message}'."
         raise wrap_error(ConnectionJdbcException, e, msg)
@@ -104,6 +113,11 @@ module LogStash module Filters module Jdbc
       ensure
         db.disconnect unless db.nil?
       end
+    end
+
+    def make_driver_path_loadable(class_loader, driver_path)
+      # so we can set an expectation in rspec
+      class_loader.add_url(java.io.File.new(driver_path).toURI().toURL())
     end
 
     def post_initialize()