+++ lib/logstash/filters/jdbc/column.rb
@@ -1,3 +1,4 @@
+# encoding: utf-8
 require_relative "validatable"
 
 module LogStash module Filters module Jdbc