+++ lib/logstash/filters/jdbc/repeating_load_runner.rb
@@ -1,3 +1,4 @@
+# encoding: utf-8
 require_relative "single_load_runner"
 
 module LogStash module Filters module Jdbc
@@ -6,6 +7,7 @@ module LogStash module Filters module Jdbc
 
     def repeated_load
       local.repopulate_all(loaders)
+      @reload_counter.increment
     end
   end
 end end end