+++ lib/logstash/filters/jdbc_static.rb
@@ -77,7 +77,7 @@ module LogStash module Filters class JdbcStatic < LogStash::Filters::Base
   #
   # There is no schedule by default. If no schedule is given, then the loaders are run
   # exactly once.
-  config :loader_schedule, :validate => [LogStash::Filters::Jdbc::LoaderSchedule]
+  config :loader_schedule, :required => false, :validate => LogStash::Filters::Jdbc::LoaderSchedule
 
   # Append values to the `tags` field if sql error occured
   # Alternatively, individual `tag_on_failure` arrays can be added to each lookup hash
@@ -87,7 +87,8 @@ module LogStash module Filters class JdbcStatic < LogStash::Filters::Base
   config :tag_on_default_use, :validate => :array, :default => ["_jdbcstaticdefaultsused"]
 
   # Remote Load DB Jdbc driver library path to third party driver library.
-  config :jdbc_driver_library, :validate => :path
+  # Use comma separated paths in one string if you need more than one library.
+  config :jdbc_driver_library, :validate => :string
 
   # Remote Load DB Jdbc driver class to load, for example "oracle.jdbc.OracleDriver" or "org.apache.derby.jdbc.ClientDriver"
   config :jdbc_driver_class, :validate => :string, :required => true
@@ -100,6 +101,9 @@ module LogStash module Filters class JdbcStatic < LogStash::Filters::Base
 
   # Remote Load DB Jdbc password
   config :jdbc_password, :validate => :password
+
+  # directory for temp files created during bulk loader import.
+  config :staging_directory, :validate => :string, :default => ::File.join(Dir.tmpdir, "logstash", config_name, "import_data")
 
   # NOTE: For the initial release, we are not allowing the user to specify their own local lookup JDBC DB settings.
   # In the near future we have to consider identical config running in multiple pipelines stomping over each other
@@ -123,6 +127,11 @@ module LogStash module Filters class JdbcStatic < LogStash::Filters::Base
         unless validation_errors.nil?
           return false, validation_errors
         end
+      elsif validator.respond_to?(:find_validation_errors)
+        validation_errors = validator.find_validation_errors(value)
+        unless validation_errors.nil?
+          return false, validation_errors
+        end
       else
         return old_validate_value(value, validator)
       end
@@ -135,7 +144,6 @@ module LogStash module Filters class JdbcStatic < LogStash::Filters::Base
   def register
     prepare_data_dir
     prepare_runner
-    @loader_runner.initial_load
   end
 
   def filter(event)
@@ -149,12 +157,18 @@ module LogStash module Filters class JdbcStatic < LogStash::Filters::Base
     @processor.close
   end
 
+  def loader_runner
+    # use for test verification
+    @loader_runner
+  end
+
   private
 
   def prepare_data_dir
     # later, when local persistent databases are allowed set this property to LS_HOME/data/jdbc-static/
     # must take multi-pipelines into account and more than one config using the same jdbc-static settings
     java.lang.System.setProperty("derby.system.home", ENV["HOME"])
+    logger.info("derby.system.home is: #{java.lang.System.getProperty("derby.system.home")}")
   end
 
   def prepare_runner
@@ -168,18 +182,20 @@ module LogStash module Filters class JdbcStatic < LogStash::Filters::Base
     @processor = Jdbc::LookupProcessor.new(@local_lookups, global_lookup_options)
     runner_args.unshift(@processor.local)
     if @loader_schedule
-      require "rufus/scheduler"
       args = []
       @loader_runner = Jdbc::RepeatingLoadRunner.new(*runner_args)
-
+      @loader_runner.initial_load
       cronline = Jdbc::LoaderSchedule.new(@loader_schedule)
+      cronline.to_log_string.tap do |msg|
+        logger.info("Scheduler operations: #{msg}") unless msg.empty?
+      end
+      logger.info("Scheduler scan for work frequency is: #{cronline.schedule_frequency}")
       rufus_args = {:max_work_threads => 1, :frequency => cronline.schedule_frequency}
-
       @scheduler = Rufus::Scheduler.new(rufus_args)
       @scheduler.cron(cronline.loader_schedule, @loader_runner)
-      @scheduler.join
     else
       @loader_runner = Jdbc::SingleLoadRunner.new(*runner_args)
+      @loader_runner.initial_load
     end
   end
 
@@ -211,6 +227,9 @@ module LogStash module Filters class JdbcStatic < LogStash::Filters::Base
     end
     if @jdbc_password
       options["jdbc_password"] = @jdbc_password
+    end
+    if @staging_directory
+      options["staging_directory"] = @staging_directory
     end
   end
 end end end