+++ spec/filters/shared_helpers.rb
@@ -12,12 +12,11 @@ RSpec.shared_examples "a single load runner" do
 
     it "builds local db objects and populates the local db" do
       expect(local_db).to receive(:populate_all).once.with(loaders)
-      expect(local_db).to receive(:build_db_object).twice.with(instance_of(LogStash::Filters::Jdbc::DbObject))
+      expect(local_db).to receive(:build_db_object).once.with(instance_of(LogStash::Filters::Jdbc::DbObject))
       runner.initial_load
       expect(runner.preloaders).to be_a(Array)
-      expect(runner.preloaders.size).to eq(2)
+      expect(runner.preloaders.size).to eq(1)
       expect(runner.preloaders[0].name).to eq(:servers)
-      expect(runner.preloaders[1].name).to eq(:temp_servers)
       expect(runner.local).to eq(local_db)
       expect(runner.loaders).to eq(loaders)
     end